linker: rename to links
