+++ tests/resources/test_pagination.py
@@ -16,7 +16,7 @@ import pytest
 from flask_principal import Identity, Need, UserNeed
 from invenio_search import current_search
 
-from invenio_records_resources.linker.schema import DEFAULT_MAX_RESULTS, \
+from invenio_records_resources.links.schema import DEFAULT_MAX_RESULTS, \
     DEFAULT_RESULTS_PER_PAGE
 from invenio_records_resources.services import RecordService
